+++ src/modules/tu/siswa/ListSiswa.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import DataTables from '../../components/data_tables/DataTables';
-import { pushLoading } from '../../components/layout/ActionLayout';
-import Modals from '../../components/modal/Modal';
+import DataTables from '../../../components/data_tables/DataTables';
+import { pushLoading } from '../../../components/layout/ActionLayout';
+import Modals from '../../../components/modal/Modal';
 import { Fade } from 'react-reveal';
 
 class ListSiswa extends Component {
